requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.0
requests==2.31.0
feedparser==6.0.10
python-dotenv==1.0.0
apscheduler==3.10.4
newspaper4k==0.9.0
lxml_html_clean==0.1.1
openai==1.0.0
textblob==0.17.1
python-dateutil==2.8.2

# WhatsApp
twilio==8.10.0

# Excel
openpyxl==3.1.2
xlsxwriter==3.1.9
pandas==2.1.3

# Web scraping
beautifulsoup4==4.12.2

# Chrome scraping (opcional - comentar si no funciona en cloud)
# undetected-chromedriver==3.5.4
# selenium==4.15.2
